refactor: decouple query read models from operations
