+++ apps/desktop/src/state/queries/workspace.ts
@@ -1,9 +1,9 @@
 import type { RepoConfig, SettingsSnapshot, WorkspaceRecord } from "@openducktor/contracts";
 import { type QueryClient, queryOptions } from "@tanstack/react-query";
+import { hostClient as host } from "@/lib/host-client";
 import { normalizeTargetBranch } from "@/lib/target-branch";
 import { DEFAULT_RUNTIME_KIND } from "@/state/agent-runtime-registry";
 import type { RepoSettingsInput } from "@/types/state-slices";
-import { host } from "../operations/host";
 
 const SETTINGS_SNAPSHOT_STALE_TIME_MS = 15 * 60_000;
 const REPO_CONFIG_STALE_TIME_MS = 10 * 60_000;
