+++ apps/desktop/src/state/read-models/task-read-model.ts
@@ -1,0 +1,4 @@
+import type { TaskCard } from "@openducktor/contracts";
+
+export const toVisibleTasks = (taskList: TaskCard[]): TaskCard[] =>
+  taskList.filter((task) => task.status !== "deferred");
